# api-spec.yaml
openapi: 3.1.0
info:
  title: MS Template API
  description: "Microservice template API exposing company, salesman, admin user,\
    \ and external posts endpoints. All endpoints support request context propagation\
    \ using X-Request-Id, X-Tenant-Id, and Accept-Language headers."
  contact:
    name: MS Template Team
    url: https://snb.com/support
    email: support@snb.com
  license:
    name: Proprietary
    url: https://snb.com/legal
  version: v1
...
servers:
- url: http://localhost:8080
  description: Local development
- url: /
  description: Relative deployment base path
...
paths:
...
  /api/rbac/privileges:
    get:
      tags:
      - RBAC
      summary: List privileges
      description: Returns all active privilege records.
      operationId: getAllPrivileges
      parameters:
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      responses:
        "200":
          description: Privileges fetched successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrivilegeResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
    post:
      tags:
      - RBAC
      summary: Create privilege
      description: Creates a new privilege record.
      operationId: createPrivilege
      parameters:
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      requestBody:
        description: Privilege payload to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivilegeCreateRequest"
            examples:
              CreatePrivilege:
                description: CreatePrivilege
                value:
                  privilegeCode: USER_VIEW
                  privilegeName: View Users
                  description: Can view user profiles
        required: true
      responses:
        "201":
          description: Privilege created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PrivilegeResponse"
        "400":
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
  /api/rbac/privileges/{privilegeCode}:
    get:
      tags:
      - RBAC
      summary: Get privilege by code
      description: Finds a privilege by code.
      operationId: getPrivilegeByCode
      parameters:
      - name: privilegeCode
        in: path
        description: Privilege code
        required: true
        schema:
          type: string
        example: USER_VIEW
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      responses:
        "200":
          description: Privilege fetched successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PrivilegeResponse"
        "400":
          description: Invalid privilege code supplied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "404":
          description: Privilege not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
    put:
      tags:
      - RBAC
      summary: Update privilege
      description: Updates an existing privilege record.
      operationId: updatePrivilege
      parameters:
      - name: privilegeCode
        in: path
        description: Privilege code
        required: true
        schema:
          type: string
        example: USER_VIEW
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      requestBody:
        description: Privilege payload to update
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivilegeUpdateRequest"
            examples:
              UpdatePrivilege:
                description: UpdatePrivilege
                value:
                  privilegeName: View All Users
                  description: Updated description
        required: true
      responses:
        "200":
          description: Privilege updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PrivilegeResponse"
        "400":
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "404":
          description: Privilege not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
    delete:
      tags:
      - RBAC
      summary: Delete privilege
      description: Soft-deletes a privilege record.
      operationId: deletePrivilege
      parameters:
      - name: privilegeCode
        in: path
        description: Privilege code
        required: true
        schema:
          type: string
        example: USER_VIEW
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      responses:
        "200":
          description: Privilege deleted successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PrivilegeResponse"
        "404":
          description: Privilege not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
...
  /api/roles/{roleCode}/privileges/{privilegeCode}:
    delete:
      tags:
      - RBAC
      summary: Revoke role-privilege grant
      description: Soft-deletes a role-privilege grant by roleCode and privilegeCode.
      operationId: revokeRolePrivilege
      parameters:
      - name: roleCode
        in: path
        description: Role code
        required: true
        schema:
          type: string
        example: SUPER_ADMIN
      - name: privilegeCode
        in: path
        description: Privilege code
        required: true
        schema:
          type: string
        example: USER_VIEW
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      responses:
        "200":
          description: Privilege revoked successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RolePrivilegeResponse"
        "404":
          description: Grant not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
...
components:
  schemas:
...
    PrivilegeCreateRequest:
      type: object
      description: Privilege creation request payload
      properties:
        privilegeCode:
          type: string
          description: Stable privilege code used in authorization checks
          example: USER_VIEW
          maxLength: 100
          minLength: 1
        privilegeName:
          type: string
          description: Human-readable privilege name
          example: View Users
          maxLength: 150
          minLength: 1
        description:
          type: string
          description: Privilege description
          example: Can view user profiles
          maxLength: 255
      required:
      - privilegeCode
      - privilegeName
    PrivilegeUpdateRequest:
      type: object
      description: Privilege update request payload
      properties:
        privilegeName:
          type: string
          description: Human-readable privilege name
          example: View All Users
          maxLength: 150
          minLength: 1
        description:
          type: string
          description: Privilege description
          example: Updated description
          maxLength: 255
      required:
      - privilegeName
    PrivilegeResponse:
      type: object
      description: Privilege response payload
      properties:
        privilegeId:
          type: integer
          format: int64
          description: Unique privilege identifier
          example: 1
        privilegeCode:
          type: string
          description: Stable privilege code used in authorization checks
          example: USER_VIEW
        privilegeName:
          type: string
          description: Human-readable privilege name
          example: View Users
        description:
          type: string
          description: Privilege description
          example: Can view user profiles
        createdBy:
          type: integer
          format: int64
          description: Identifier of actor who created the record
          example: 1001
        createdAt:
          type: string
          format: date-time
          description: Record creation timestamp in ISO-8601 format
          example: 2026-05-04T10:15:30
        updatedBy:
          type: integer
          format: int64
          description: Identifier of actor who last updated the record
          example: 1002
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp in ISO-8601 format
          example: 2026-05-04T11:20:45
...
    RolePrivilegeResponse:
      type: object
      description: Role-privilege grant response payload
      properties:
        id:
          type: integer
          format: int64
          description: Grant record identifier
          example: 1
        roleId:
          type: integer
          format: int64
          description: Role identifier
          example: 1
        roleCode:
          type: string
          description: Role code
          example: SUPER_ADMIN
        privilegeId:
          type: integer
          format: int64
          description: Privilege identifier
          example: 1
        privilegeCode:
          type: string
          description: Privilege code
          example: USER_VIEW
        privilegeName:
          type: string
          description: Privilege name
          example: View Users
        createdBy:
          type: integer
          format: int64
          description: Identifier of actor who created the record
          example: 1
        createdAt:
          type: string
          format: date-time
          description: Record creation timestamp
          example: 2026-05-19T10:00:00
        updatedBy:
          type: integer
          format: int64
          description: Identifier of actor who last updated the record
          example: 1
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: 2026-05-19T11:00:00
...
    ErrorInfo:
      type: object
      description: Detailed error information
      properties:
        type:
          type: string
          description: Error category
          example: VALIDATION_ERROR
        code:
          type: string
          description: Application-specific error code
          example: EMAILADDRESS_INVALID
        message:
          type: string
          description: "Localized, user-friendly error message"
          example: emailAddress must be a valid email
        description:
          type: string
          description: Additional diagnostic details
          example: "Rejected value: not-an-email"
    BaseResponseDTO:
      type: object
      description: Base API response wrapper with error details
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorInfo"
...
  parameters:
    XRequestIdHeader:
      name: X-Request-Id
      in: header
      description: "Correlation identifier for distributed tracing. If omitted, the\
        \ service generates one."
      required: false
      schema:
        type: string
        example: req-20260504-0001
    XTenantIdHeader:
      name: X-Tenant-Id
      in: header
      description: Tenant identifier used for multi-tenant context propagation.
      required: false
      schema:
        type: string
        example: tenant-001
    AcceptLanguageHeader:
      name: Accept-Language
      in: header
      description: Preferred locale for localized validation and error messages.
      required: false
      schema:
        type: string
        example: en